[Protocol] Add common code file
