+++ src/protocols/gamespy/protocols/one/protocol.rs
@@ -9,6 +9,7 @@ use crate::{
     GDResult,
 };
 
+use crate::protocols::gamespy::common::has_password;
 use std::collections::HashMap;
 
 fn get_server_values(
@@ -171,21 +172,6 @@ fn extract_players(server_vars: &mut HashMap<String, String>, players_maximum: u
     }
 
     Ok(players)
-}
-
-fn has_password(server_vars: &mut HashMap<String, String>) -> GDResult<bool> {
-    let password_value = server_vars
-        .remove("password")
-        .ok_or(GDError::PacketBad)?
-        .to_lowercase();
-
-    if let Ok(has) = password_value.parse::<bool>() {
-        return Ok(has);
-    }
-
-    let as_numeral: u8 = password_value.parse().map_err(|_| GDError::TypeParse)?;
-
-    Ok(as_numeral != 0)
 }
 
 /// If there are parsing problems using the `query` function, you can directly
