+++ src/protocols/gamespy/protocols/three/protocol.rs
@@ -1,4 +1,5 @@
 use crate::bufferer::{Bufferer, Endianess};
+use crate::protocols::gamespy::common::has_password;
 use crate::protocols::gamespy::three::{Player, Response};
 use crate::protocols::types::TimeoutSettings;
 use crate::socket::{Socket, UdpSocket};
@@ -176,21 +177,6 @@ pub fn query_vars(
     }
 
     Ok(vars)
-}
-
-fn has_password(server_vars: &mut HashMap<String, String>) -> GDResult<bool> {
-    let password_value = server_vars
-        .remove("password")
-        .ok_or(GDError::PacketBad)?
-        .to_lowercase();
-
-    if let Ok(has) = password_value.parse::<bool>() {
-        return Ok(has);
-    }
-
-    let as_numeral: u8 = password_value.parse().map_err(|_| GDError::TypeParse)?;
-
-    Ok(as_numeral != 0)
 }
 
 fn parse_parse_players(packets: Vec<Vec<u8>>) -> GDResult<Vec<Player>> {
